Decompiled func_8003671C

+++ include/code0/37090.h
@@ -6,7 +6,7 @@
 s32 func_80036490(s16 sectnum);
 void func_80036520(s16 playernum, s16);
 void func_800365C0(s16);
-void func_8003671C(s16 playernum, u16, s32, s32);
+void func_8003671C(s16 playernum, u16, s16, s16);
 s16 func_8003B31C(s16 playernum, s16);
 void func_8003DACC(void);
 void func_8003DD54(void);
